requirements.txt
# Core framework
pathway[all]>=0.7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP and networking
requests>=2.31.0
httpx>=0.25.0

# Environment and configuration
python-dotenv>=1.0.0

# AI and ML
openai>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# Data processing
pandas>=2.1.0

# Web server utilities
aiofiles>=23.2.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Utilities
geopy>=2.4.0
schedule>=1.2.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
